Implement GH issue to PT story task hook
AddPtTaskOnGhIssue method of Workflow implements the specified hook. The
hook listens for github.issues events. When an event of this kind is
received, the issue body is check for Pivotal Tracker story URL. In case
a URL is found, a task is added to the relevant PT story containing the
link to the source GitHub issue.